# [Internal] : Change LazyCosmosObject's cachedElement property from Dictionary to Lazy<Dictionary>