# Pull Request Template

## Description

Change initialization of cached elements to Lazy<>, to improve managed memory

## Type of change

Please delete options that are not relevant.

- [] Bug fix (non-breaking change which fixes an issue)


## Closing issues

To automatically close an issue: closes #IssueNumber