@darshdeeph 2020-10-08
> Please submit a benchmark that shows the memory improvement. We can also just remove the caching layer and have the caller wrap the object.

I checked improvement by creating dump file while running an integration test locally on visual studio and checking the object allocation there. Does that suffice, or do we need something more involved?