@darshdeeph 2020-10-15
With this change, LazyCosmosElementsBenchmarks

                         Method |     payload | jsonSerializationFormat | accessMode |        Mean |     Error |  Gen 0 | Allocated
 CreateAndNavigateCosmosElement |      object |                    Text |      Never |    148.5 ns |   1.79 ns | 0.0777 | 488 B
 CreateAndNavigateCosmosElement |      object |                    Text |       Once |  9,605.4 ns |  56.46 ns | 0.9766 | 6176 B
 CreateAndNavigateCosmosElement |      object |                    Text |       Many | 39,295.3 ns | 223.86 ns | 6.6528 | 41960 B
 CreateAndNavigateCosmosElement |      object |                  Binary |      Never |    135.7 ns |   0.67 ns | 0.0548 | 344 B
 CreateAndNavigateCosmosElement |      object |                  Binary |       Once |  6,146.8 ns |  24.49 ns | 1.0147 | 6376 B
 CreateAndNavigateCosmosElement |      object |                  Binary |       Many | 59,887.0 ns | 477.80 ns | 9.6436 | 60664 B


Without this change, LazyCosmosElementsBenchmarks

                         Method |     payload | jsonSerializationFormat | accessMode |        Mean |     Error  |  Gen 0 | Allocated
 CreateAndNavigateCosmosElement |      object |                    Text |      Never |    250.7 ns |   1.74 ns | 0.1988 |   1248 B
 CreateAndNavigateCosmosElement |      object |                    Text |       Once |  9,829.9 ns |  62.93 ns | 1.2207 |  7696 B
 CreateAndNavigateCosmosElement |      object |                    Text |       Many | 40,031.1 ns | 178.95 ns | 7.9956 |   50320 B
 CreateAndNavigateCosmosElement |      object |                  Binary |      Never |    223.1 ns |   1.29 ns | 0.1760 |    1104 B
 CreateAndNavigateCosmosElement |      object |                  Binary |       Once |  6,302.9 ns |  32.10 ns | 1.2512 |   7896 B
 CreateAndNavigateCosmosElement |      object |                  Binary |       Many | 60,968.6 ns | 253.21 ns |10.9863 |  69024 B

We can see that the the mean latency and allocation are extremely lower for Never access mode, and slightly lower for other access modes.